Undo/Redo 70%

+++ GeRuMap/src/main/java/dsw/raf/geruMap/StateManager/States/MoveState.java
@@ -6,6 +6,8 @@ import dsw.raf.geruMap.MapRepository.Implementation.Link;
 import dsw.raf.geruMap.MapRepository.Implementation.Thought;
 import dsw.raf.geruMap.MapRepository.MapRepositoryImpl;
 import dsw.raf.geruMap.MapRepository.Painters.ElementPainter;
+import dsw.raf.geruMap.MapRepository.Painters.LinkPainter;
+import dsw.raf.geruMap.gui.swing.commands.implementation.MoveCommand;
 import dsw.raf.geruMap.gui.swing.tree.model.MapTreeItem;
 import dsw.raf.geruMap.gui.swing.view.MainFrame;
 import dsw.raf.geruMap.gui.swing.view.MapView;
@@ -15,6 +17,7 @@ import java.awt.event.MouseWheelEvent;
 
 public class MoveState implements AbstractState{
     Point start;
+    Point origin;
     boolean single;
 
     @Override
@@ -60,17 +63,35 @@ public class MoveState implements AbstractState{
             for (ElementPainter i : ((MapView) MainFrame.getInstance().getDesktop().getSelectedComponent()).getElems())
             {
                 if (i.elementAt(i.getElement(),new Point(x,y)))
+                {
                     ((MapRepositoryImpl)AppCore.getInstance().getRep()).getMapSelection().add_selected(i.getElement());
 
+                    if (i instanceof LinkPainter)
+                    {
+                        Link temp = (Link) i.getElement();
+
+                        ((MapRepositoryImpl)AppCore.getInstance().getRep()).getMapSelection().add_selected(temp.getParentThought());
+                        ((MapRepositoryImpl)AppCore.getInstance().getRep()).getMapSelection().add_selected(temp.getChildThought());
+                    }
+                }
             }
         }
 
         start = new Point(x,y);
+        origin = start;
     }
 
     @Override
     public void mouseRelease(int x, int y)
     {
+        if (((MapRepositoryImpl)AppCore.getInstance().getRep()).getMapSelection().getSelection().isEmpty())
+            return;
+
+        Point delta = new Point(start.x - origin.x, start.y - origin.y);
+
+        AppCore.getInstance().getGui().getCommandManager().addCommand(new MoveCommand(((MapRepositoryImpl)AppCore.getInstance().getRep()).getMapSelection().getSelection(), delta));
+        AppCore.getInstance().getGui().getCommandManager().doCommand();
+
         if (single)
             ((MapRepositoryImpl)AppCore.getInstance().getRep()).getMapSelection().remove_all();
         single = false;
